Move AccountDropdown into its own folder & pull out children components
