+++ ui/src/AccountDropdown/AccountDropdown.scss
@@ -43,31 +43,28 @@
   }
 }
 
-.account-dropdown-options {
-  text-align: left;
-  padding: 4px 6px;
-}
-
-.account-dropdown-options:hover {
-  background-color: $gray-6;
-  color: $purple;
-  border-radius: 2px;
-}
-
-.dropdown-container {
+.accountDropdown {
   position: absolute;
   background-color: white;
   width: 150px;
   border-radius: 6px;
   border: 1px solid #EDEBEB;
-
   right: 12px;
   top: 40px;
-
   box-shadow: 1px 2px 3px rgba(0, 0, 0, .3);
-
   padding: 16px 15px;
   box-sizing: border-box;
   font-family: Helvetica, sans-serif;
   font-size: 12px;
+
+  .accountDropdownOption {
+    text-align: left;
+    padding: 4px 6px;
+  }
+
+  .accountDropdownOption:hover {
+    background-color: $gray-6;
+    color: $purple;
+    border-radius: 2px;
+  }
 }
